separated boundaries CPU functions
